move hook theme in root
