--- src/components/hooks/index.js
@@ -1,3 +1,0 @@
-import useTheme from "./useTheme";
-
-export { useTheme };
